Added loaders for the news and user profile pages, and made minor layout adjustments
