+++ src/Components/Loading/Loading.jsx
@@ -1,0 +1,5 @@
+import styles from "./styles.module.css";
+
+export default function Loading() {
+  return <div className={styles.loading}></div>;
+}
