+++ src/pages/HomePage/styles.module.css
@@ -9,6 +9,7 @@
 
 .title {
   font-size: 20px;
+  text-align: center;
 }
 
 .nav {
